refactor: Extract store
